+++ apps/ui-dev/component/MarkdownPage.jsx
@@ -5,10 +5,10 @@ export function MarkdownPage({ page }) {
   const template = page();
 
   const items = [];
-  template.children.forEach((child) => {
+  template.children.forEach((child, index) => {
     if (child.nodeName?.startsWith("H")) {
       const text = child.makeText(" ");
-      const id = child.makeText("-");
+      const id = child.makeText("-") + index;
       child.props.id = encodeURIComponent(id);
       child.tagProps.id = encodeURIComponent(id);