+++ apps/ui-dev/component/TableOfContents.jsx
@@ -3,6 +3,7 @@ export function TableOfContents({ items }) {
   return (
     <div class="table-of-contents">
       <div class="items">
+        <div class="title">Table of contents</div>
         {items.map((it) => {
           return (
             <div class="toc-item" data-level={it.level}>